daemon/libnetwork: move unmarshalJSONField to json_util.go
unmarshalJSONField is called from endpoint.go, endpoint_info.go, and
network.go, so keeping it (and its test) in endpoint.go made it
harder to find. Move it to its own file.
